requirements.txt
flask==0.12   # Known XSS vulnerability
pyjwt==1.7.1  # JWT signature bypass vulnerability
